% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_outcome.R
\name{asymptotic_var_general_binary}
\alias{asymptotic_var_general_binary}
\title{Asymptotic variance based on sGIM for the binary outcome}
\usage{
asymptotic_var_general_binary(beta_est,gamma_est,theta_est,
v_ext,y_int,x_int,used_index)
}
\arguments{
\item{beta_est}{the estimates of the main parameter vector}

\item{gamma_est}{the estimates of the shift parameter}

\item{theta_est}{the estimates of parameters in the reduced model}

\item{v_ext}{the variance-covariance matrix from the external source}

\item{y_int}{the outcome vector from the internal data}

\item{x_int}{the covariate matrix from the internal data}

\item{used_index}{index for parameters in the reduced model in relation to the full model}
}
\value{
The asymptotic variance-covariance matrix
}
\description{
This function calculate asymptotic variance of sGIM
}
