% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_outcome.R
\name{information.borrowing_lgim_binary}
\alias{information.borrowing_lgim_binary}
\title{sGIM for the binary outcome}
\usage{
information.borrowing_lgim_binary(beta_initial,gamma_initial,
theta_ext_ini,theta_ext_est,v_ext,y_int,x_int,used_index)
}
\arguments{
\item{beta_initial}{the initial values of the main parameter vector}

\item{gamma_initial}{the initial values of the shift parameter}

\item{theta_ext_ini}{the initial values of parameters in the reduced model}

\item{theta_ext_est}{the summary information (parameter estimates) from the external source}

\item{v_ext}{the variance-covariance matrix from the external source}

\item{y_int}{the outcome vector from the internal data}

\item{x_int}{the covariate matrix from the internal data}

\item{used_index}{index for parameters in the reduced model in relation to the full model}
}
\value{
A list of parameter estimates
}
\description{
This function can integrate summary information from external source
}
